Fix compilation errors with msvc.
1. Our version of msvc can't compile std::isnan(a) where `a` is a uint8_t.
This is a bug per https://en.cppreference.com/w/cpp/numeric/math/isnan.  Maybe
it's fixed in newer versions. :shrug:

2. Our version of msvc can't handle the large m::Tuple() expressions introduced
in algebraic_simplifier_test.  Write the test a different way.

PiperOrigin-RevId: 416968536

+++ tensorflow/compiler/xla/literal.cc
@@ -1785,7 +1785,11 @@ bool LiteralBase::operator==(const LiteralBase& other) const {
 
 template <typename NativeT>
 static bool EqualIncludingNan(NativeT a, NativeT b) {
-  return a == b || (std::isnan(a) && std::isnan(b));
+  // msvc can't compile std::isnan(a) where `a` is uint8_t.  This is a bug
+  // according to https://en.cppreference.com/w/cpp/numeric/math/isnan, but it's
+  // easy to work around.
+  return a == b || (std::isnan(static_cast<double>(a)) &&
+                    std::isnan(static_cast<double>(b)));
 }
 
 template <typename T>
